change of background-index dir, to static/css, and fix to flash messages

+++ app.py
@@ -125,6 +125,9 @@ def update(id):
         flash('No new Data to Update', 'info')
         return render_template('search.html', movie=movie_update)
     movie_update_req = request.files['file']
+    if not allowed_file(movie_update_req.filename):
+        flash('Invalid file type', 'info')
+        return render_template('search.html', movie=movie_update)
     if 'file' in request.files and allowed_file(movie_update_req.filename):
         os.remove(os.path.join(app.config['UPLOAD_FOLDER'], movie_update.file))
         file_name = secure_filename(movie_update_req.filename)
